fix: rename stepperPageHeader
