+++ src/App.tsx
@@ -1,12 +1,9 @@
 import { Route, Routes } from "react-router-dom"
-import { Portfolio } from "./pages"
-import { AddPortfolis } from "./pages/AddPortfolio"
+import { AddPortfolis } from "./pages"
 
 function App() {
-
   return (
       <Routes>
-         {/* <Route path='/' element={<Portfolio/>}/> */}
          <Route path='/' element={<AddPortfolis/>}/>
       </Routes>
   )
