+++ src/components/index.ts
@@ -2,4 +2,5 @@ export * from './Button';
 export * from './Header';
 export * from './SelectBox';
 export * from './languageSelectbox';
-export * from './SearchBar'
+export * from './SearchBar';
+export * from './StepperPageHeader'
