+++ src/pages/index.ts
@@ -1,1 +1,2 @@
 export * from './Portfolios';
+export * from './AddPortfolio'
